requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.8.24
Click==7.0
colorama==0.4.3
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
h5py==2.10.0
html5lib==0.9999999
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
matplotlib==3.0.3
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.18.0
object-detection==0.1
opencv-python==4.1.2.30
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
zipp==0.6.0
